fix: fix types for eslint

+++ packages/cli/src/lib/commands/untrack.ts
@@ -1,21 +1,26 @@
 import { getIdMapperClientByName } from '../services';
 import { Container } from 'typedi';
-import { LOGGER } from '../constants';
 import pino from 'pino';
+import { LOGGER } from '../constants';
 
-export async function runUntrack({
-  collection,
-  id,
-}: {
+interface RunUntrackOptions {
   collection: string;
   id: string;
-}) {
-  const logger = Container.get(LOGGER) ;
+}
+
+export async function runUntrack(options?: RunUntrackOptions) {
+  const logger: pino.Logger = Container.get(LOGGER);
+  if (!options) {
+    throw new Error('Missing options');
+  }
+  const { collection, id } = options;
   const idMapper = getIdMapperClientByName(collection);
   try {
     await idMapper.removeByLocalId(id);
     logger.info(`Untracked ${collection} ${id}`);
   } catch (e) {
-    logger.warn(`Failed to untrack ${collection} ${id}: ${e}`);
+    logger.warn(
+      `Failed to untrack ${collection} ${id}: ${(e as Error).message}`,
+    );
   }
 }
